requirements.txt
numpy
tqdm
zhon
Pillow
PyYAML
pypinyin
soundfile
dashscope
librosa
moviepy==1.0.3
opencv-python
torch
transformers
diffusers
accelerate
edge_tts
openai
git+https://github.com/aliyun/alibabacloud-nls-python-sdk@dev